backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
SQLAlchemy>=2.0.0
PyJWT>=2.0.0
python-jose[cryptography]==3.3.0
